filtration and handling logic moved to utils/fileHandler.js
